Hotfix - Fixed master don't unregister server after abrupt wrapper shutdown - Fixed PoloCloud-API.jar file won't be copied correctly

+++ polocloud-wrapper/src/main/java/de/polocloud/wrapper/manager/server/ServiceStarter.java
@@ -171,12 +171,17 @@ public class ServiceStarter {
             FileUtils.copyDirectory(FileConstants.WRAPPER_EVERY_MINECRAFT_TEMPLATE, serverLocation); //Every spigot template
         }
 
-        try {
-            URL inputUrl = getClass().getResource("/" + FileConstants.CLOUD_API_NAME);
-            FileUtils.copyURLToFile(inputUrl, new File(serverLocation + "/plugins/" + FileConstants.CLOUD_API_NAME));
-            return;
-        } catch (Exception e) {
-            e.printStackTrace();
+
+        if(new File(FileConstants.WRAPPER_TEMPLATES + "/" + FileConstants.CLOUD_API_NAME).exists()){
+            FileUtils.copyFile(new File(FileConstants.WRAPPER_TEMPLATES + "/" + FileConstants.CLOUD_API_NAME), new File(serverLocation + "/plugins/" + FileConstants.CLOUD_API_NAME));
+        }else{
+            try {
+                URL inputUrl = getClass().getResource("/" + FileConstants.CLOUD_API_NAME);
+                FileUtils.copyURLToFile(inputUrl, new File(serverLocation + "/plugins/" + FileConstants.CLOUD_API_NAME));
+                return;
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
         }
 
         File serverIcon = new File(serverLocation, "server-icon.png");
